Improve structure and API (#295)
Description
===========

This is a first cleanup patch to bring the API shape and form
more in line with our other newer plugins. This means we use
`gradle-commons` and `gradle-commons-test` helper libraries as
well as reducing the property API to a minimum to help with testing
and to not overload the API too much. Some of the changes are breaking
but it is a good time to do these cleanups.

Changes
=======

* ![ADD] `gradle-commons` helper library for plugin runtime
* ![Add] `gradle-commons-test` helper library for tests
* ![REMOVE] certain overloaded APIs
* ![ADD] `BaseSpec` trait object
* ![IMPROVE] make Interfaces trait objects