Revert secretResolver setter in secrets extension (#298)
Description
===========

In patch [#294] I removed a setter in the secrets extension to
set the secret resolver. Since we use a default resolver chain I thought
we don't need this setter anymore. I checked with most of the projects
implementing this plugin and I saw that this setter is used a lot and
would actually make upgrade a lot harder. In the end I decided to revert
this decision and add it in again. It will basically clear the chain
and set a single resolver. That was from the old behavior the same so
nothing actually changes here in API usage

Changes
=======

* ![ADD] `setSecretResolver` setter in secrets extension